requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
chromadb
pdf2image
faiss-cpu
langchain_google_genai